refactor: move /internal/browser package to /pkg/browser
